big work done xD

+++ code/renderervk/tr_image.cpp
@@ -233,7 +233,7 @@ void R_SetColorMappings()
 
 	shift = tr.overbrightBits;
 
-	for (i = 0; i < arrayLen(s_gammatable); i++)
+	for (i = 0; i < static_cast<int>(arrayLen(s_gammatable)); i++)
 	{
 		if (g == 1.0f)
 		{
@@ -255,7 +255,7 @@ void R_SetColorMappings()
 		s_gammatable[i] = inf;
 	}
 
-	for (i = 0; i < arrayLen(s_intensitytable); i++)
+	for (i = 0; i < static_cast<int>(arrayLen(s_intensitytable)); i++)
 	{
 		j = i * r_intensity->value;
 		if (j > 255)
@@ -347,7 +347,7 @@ void TextureMode(const char *string)
 	int i;
 
 	mode = NULL;
-	for (i = 0; i < arrayLen(modes); i++)
+	for (i = 0; i < static_cast<int>(arrayLen(modes)); i++)
 	{
 		if (!Q_stricmp(modes[i].name, string))
 		{
@@ -548,7 +548,7 @@ static void ResampleTexture(unsigned *in, int inwidth, int inheight, unsigned *o
 	unsigned p2[MAX_TEXTURE_SIZE];
 	byte *pix1, *pix2, *pix3, *pix4;
 
-	if (outwidth > arrayLen(p1))
+	if (outwidth > static_cast<int>(arrayLen(p1)))
 		ri.Error(ERR_DROP, "ResampleTexture: max width");
 
 	fracstep = inwidth * 0x10000 / outwidth;
